requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
pandas==2.2.1
numpy==1.26.4
plotly==5.18.0
python-dotenv==1.0.1
Werkzeug==3.0.1
email-validator==2.1.1
Flask-Admin==1.6.1
Flask-Mail==0.9.1 